Add menu and fullscreen option
