+++ src/scss/menu.scss
@@ -1,0 +1,89 @@
+@import './variables.scss';
+
+.menu {
+    left: 1rem;
+    display: flex;
+    flex-wrap: wrap;
+    position: absolute;
+    top: 1rem;
+    z-index: 2;
+
+    .toggle { // <button>
+        background-color: transparent;
+        margin: 0rem;
+        padding: 0rem;
+
+        &:focus-visible {
+            outline: none;
+        }
+
+        &.open {
+            
+            label {
+                border-bottom-left-radius: 0rem;
+                border-bottom-right-radius: 0rem;
+                border-bottom: none;
+            }
+        }
+        
+        label {
+            cursor: pointer;
+            align-items: center;
+            border: 0.125rem solid #fff;
+            border-radius: 1rem;
+            box-shadow: inset 0.125rem 0.125rem rgba(#000, 0.2), 0.125rem 0.125rem rgba(#000, 0.2);
+            color: #ffffff;
+            display: flex;
+            font-size: 0.75rem !important;
+            font-weight: bold;
+            column-gap: 0.5rem;
+            margin: 0;
+            text-shadow: 0.0625rem 0.0625rem rgba(#000, 0.2) !important;
+            padding: 0.125rem 0.5rem !important;
+        }
+    }
+
+    .nav { // <ul>
+        background-color: $color-primary;
+        border: 0.125rem solid #fff;
+        border-radius: 1rem;
+        border-top-left-radius: 0rem;
+        box-shadow: inset 0.125rem 0.125rem rgba(#000, 0.2), 0.125rem 0.125rem rgba(#000, 0.2);
+        display: none;
+        list-style: none;
+        margin: 0rem;
+        padding: 0.5rem;
+        width: 100%;
+
+        &.visible {
+            display: block;
+        }
+
+        li {
+            button {
+                align-items: center;
+                background-color: transparent;
+                padding: 0;
+                display: flex;
+                column-gap: 0.5rem;
+                color: #fff;
+                font-size: 0.75rem;
+                font-weight: bold;
+                margin: 0;
+                text-shadow: 0.0625rem 0.0625rem rgba(#000, 0.2) !important;
+                width: 100%;
+
+                .material-symbols-outlined {
+                    border-radius: 99rem;
+                    transition: all 0.15s ease-in-out;
+                }
+
+                &:hover {
+                    .material-symbols-outlined {
+                        background-color: rgba(#000, 0.2);
+                    }
+                }
+            }
+        }
+    }
+}
