+++ src/scss/shop.scss
@@ -26,11 +26,11 @@
 			white-space: nowrap;
 
 			&:hover {
-				border-color: #ff5722;
+				border-color: $color-primary;
 			}
 			
 			&.toggle {
-				background-color: #ff5722;
+				background-color: $color-primary;
 				text-shadow: 0.125rem 0.125rem rgba(#000, 0.2);
 				box-shadow: inset 0.125rem 0.125rem rgba(#000, 0.2), 0.125rem 0.125rem rgba(#000, 0.2);
 				color: #fff;
@@ -81,7 +81,7 @@
 					}
 
 					.icon {
-						background-color: #ff5722;
+						background-color: $color-primary;
 						animation: wiggle 0.5s ease-in-out;
 					}
 
@@ -163,7 +163,7 @@
 				.description {
 					box-shadow: inset 0.125rem 0.125rem rgba(#000, 0.2), 0.125rem 0.125rem rgba(#000, 0.2);
 					border: 0.125rem solid #ffffff;
-					border-radius: 1rem;
+					border-radius: 1.5rem;
 					display: flex;
 					flex-direction: column;
 					align-items: center;
