+++ src/scss/style.scss
@@ -45,7 +45,7 @@ html {
 				flex-grow: 1;
 			}
 
-			.menu {
+			.aside {
 				height: inherit;
 				max-width: 50%;
 				width: 100%;
