--- src/scss/version.scss
@@ -1,15 +1,0 @@
-@import './variables.scss';
-
-.version {
-    border: 0.125rem solid #fff;
-    box-shadow: inset 0.125rem 0.125rem rgba(#000, 0.2), 0.125rem 0.125rem rgba(#000, 0.2);
-    color: #ffffff;
-    font-size: 0.75rem !important;
-    font-weight: bold;
-    text-shadow: 0.0625rem 0.0625rem rgba(#000, 0.2) !important;
-    left: 1rem;
-    padding: 0.125rem 0.5rem !important;
-    position: absolute;
-    top: 1rem;
-    z-index: 2;
-}
